renamed `_test_data` to `_sample_data`
