+++ src/napari_stl_exporter/__init__.py
@@ -2,4 +2,4 @@ __version__ = "0.0.10"
 
 from ._writer import napari_write_labels, napari_write_surfaces
 from ._reader import napari_import_surface
-from ._test_data import make_pyramid_label, make_pyramid_surface
+from ._sample_data import make_pyramid_label, make_pyramid_surface
