+++ src/napari_stl_exporter/_sample_data.py
